Make nicer example
