+++ pgx-examples/composite_type/src/lib.rs
@@ -6,15 +6,40 @@ All rights reserved.
 
 Use of this source code is governed by the MIT license that can be found in the LICENSE file.
 */
+
+/** A guided example of composite types in `pgx`
+
+Composite types in `pgx` are essentially macros that alias to a [`PgHeapTuple`][pgx::PgHeapTuple]
+with one important difference:
+
+During SQL generation, the composite type name provided to the [`pgx::composite_type`][pgx::composite_type]
+macro is used as the type name, whereas [`PgHeapTuple`][pgx::PgHeapTuple] does not have a valid
+Rust to SQL mapping.
+
+A number of considerations are detailed in the [`pgx::composite_type`][pgx::composite_type] and
+[`PgHeapTuple`][pgx::PgHeapTuple] documentation which may assist in productionalizing extensions
+using composite types.
+*/
+
 use pgx::*;
 
+// All `pgx` extensions will do this:
 pg_module_magic!();
 
-// This is used by some, but not all, examples below.
-const DOG_COMPOSITE_TYPE_IDENT: &str = "Dog";
+/* Composite types must be defined either before they are used.
 
-extension_sql!(
-    r#"
+This means composite types must:
+* Be defined prior to loading the extension
+* Be defined by an `extension_sql!()` block which the function has been "ordered before" in
+    the generated SQL
+
+If your extension depends on already existing composite types, then you don't need to do
+anything to define them. Just use them as if they exist.
+
+If your extension defines the composite types itself, it's recommended to do that in an
+`extension_sql!()` which is set to be a `bootstrap`, and is ordered first in the generated SQL:
+*/
+extension_sql!("\
 CREATE TYPE Dog AS (
     name TEXT,
     scritches INT
@@ -23,427 +48,95 @@ CREATE TYPE Dog AS (
 CREATE TYPE Cat AS (
     name TEXT,
     boops INT
-);
-
-CREATE TYPE Fish AS (
-    name TEXT,
-    bloops INT
-);
-
-CREATE TYPE AnimalFriendshipEdge AS (
-    friend_1_name TEXT,
-    friend_2_name TEXT
-);
-"#,
-    name = "create_dog",
+);",
+    name = "create_composites",
     bootstrap
 );
 
-// As Arguments
-mod arguments {
-    use super::*;
+/*
+If it's required for exotic reasons that the composite type not be part of the bootstrap code,
+it can also be defined in a non-bootstrap SQL block (**Note:** This could have just been
+included in the above bootstrap code):
+*/
+extension_sql!("\
+CREATE TYPE CatAndDogFriendship AS (
+    cat Cat,
+    dog Dog
+);",
+    name = "create_cat_and_dog_friendship",
+);
 
-    mod singleton {
-        use super::*;
 
-        #[pg_extern]
-        fn gets_name_field(dog: Option<pgx::composite_type!("Dog")>) -> Option<&str> {
-            // Gets resolved to:
-            let dog: Option<PgHeapTuple<AllocatedByRust>> = dog;
+// To assist with code reuse, consider setting your composite type names in constants:
+const DOG_COMPOSITE_TYPE: &str = "Dog";
+const CAT_COMPOSITE_TYPE: &str = "Cat";
+const CAT_AND_DOG_FRIENDSHIP_COMPOSITE_TYPE: &str = "CatAndDogFriendship";
+/*
+While it may be tempting to do something like:
 
-            dog?.get_by_name("name").ok()?
-        }
+```rust
+const CAT_COMPOSITE_TYPE: PgHeapTuple<'static, AllocatedByRust> = pgx::composite_type!("Cat");
+```
+This is not correct! The [`#[pgx::pg_extern]`][pgx::pg_extern] macro would not be able to pick
+up the required metadata for SQL generation. Instead, set the constant as a `&'static str` which
+the [`pgx::composite_type!()`][pgx::composite_type] macro can consume.
+*/
 
-        #[pg_extern]
-        fn gets_name_field_default(
-            dog: default!(pgx::composite_type!("Dog"), "ROW('Nami', 0)::Dog"),
-        ) -> &str {
-            // Gets resolved to:
-            let dog: PgHeapTuple<AllocatedByRust> = dog;
 
-            dog.get_by_name("name").unwrap().unwrap()
-        }
+/*
+Create a dog inside Rust then return it, roughly the equivalent of this SQL:
 
-        #[pg_extern]
-        fn gets_name_field_strict(dog: pgx::composite_type!("Dog")) -> &str {
-            // Gets resolved to:
-            let dog: PgHeapTuple<AllocatedByRust> = dog;
-
-            dog.get_by_name("name").unwrap().unwrap()
-        }
-    }
-
-    mod variadic_array {
-        use super::*;
-
-        #[pg_extern]
-        fn gets_name_field_variadic(
-            dogs: VariadicArray<pgx::composite_type!(DOG_COMPOSITE_TYPE_IDENT)>,
-        ) -> Vec<String> {
-            // Gets resolved to:
-            let dogs: pgx::VariadicArray<PgHeapTuple<AllocatedByRust>> = dogs;
-
-            let mut names = Vec::with_capacity(dogs.len());
-            for dog in dogs {
-                let dog = dog.unwrap();
-                let name = dog.get_by_name("name").unwrap().unwrap();
-                names.push(name);
-            }
-            names
-        }
-
-        #[pg_extern]
-        fn gets_name_field_default_variadic(
-            dogs: default!(
-                VariadicArray<pgx::composite_type!("Dog")>,
-                "ARRAY[ROW('Nami', 0)]::Dog[]"
-            ),
-        ) -> Vec<String> {
-            // Gets resolved to:
-            let dogs: pgx::VariadicArray<PgHeapTuple<AllocatedByRust>> = dogs;
-
-            let mut names = Vec::with_capacity(dogs.len());
-            for dog in dogs {
-                let dog = dog.unwrap();
-                let name = dog.get_by_name("name").unwrap().unwrap();
-                names.push(name);
-            }
-            names
-        }
-
-        #[pg_extern]
-        fn gets_name_field_strict_variadic(
-            dogs: pgx::VariadicArray<pgx::composite_type!("Dog")>,
-        ) -> Vec<String> {
-            // Gets resolved to:
-            let dogs: pgx::VariadicArray<PgHeapTuple<AllocatedByRust>> = dogs;
-
-            let mut names = Vec::with_capacity(dogs.len());
-            for dog in dogs {
-                let dog = dog.unwrap();
-                let name = dog.get_by_name("name").unwrap().unwrap();
-                names.push(name);
-            }
-            names
-        }
-    }
-
-    mod vec {
-        use super::*;
-
-        #[pg_extern]
-        fn sum_scritches_for_names(dogs: Option<Vec<pgx::composite_type!(DOG_COMPOSITE_TYPE_IDENT)>>) -> i32 {
-            // Gets resolved to:
-            let dogs: Option<Vec<PgHeapTuple<AllocatedByRust>>> = dogs;
-
-            let dogs = dogs.unwrap();
-            let mut sum_scritches = 0;
-            for dog in dogs {
-                let scritches: i32 = dog
-                    .get_by_name("scritches")
-                    .ok()
-                    .unwrap_or_default()
-                    .unwrap_or_default();
-                sum_scritches += scritches;
-            }
-            sum_scritches
-        }
-
-        #[pg_extern]
-        fn sum_scritches_for_names_default(
-            dogs: pgx::default!(
-                Vec<pgx::composite_type!("Dog")>,
-                "ARRAY[ROW('Nami', 0)]::Dog[]"
-            ),
-        ) -> i32 {
-            // Gets resolved to:
-            let dogs: Vec<PgHeapTuple<AllocatedByRust>> = dogs;
-
-            let mut sum_scritches = 0;
-            for dog in dogs {
-                let scritches: i32 = dog
-                    .get_by_name("scritches")
-                    .ok()
-                    .unwrap_or_default()
-                    .unwrap_or_default();
-                sum_scritches += scritches;
-            }
-            sum_scritches
-        }
-
-        #[pg_extern]
-        fn sum_scritches_for_names_strict(dogs: Vec<pgx::composite_type!("Dog")>) -> i32 {
-            // Gets resolved to:
-            let dogs: Vec<PgHeapTuple<AllocatedByRust>> = dogs;
-
-            let mut sum_scritches = 0;
-            for dog in dogs {
-                let scritches: i32 = dog
-                    .get_by_name("scritches")
-                    .ok()
-                    .unwrap_or_default()
-                    .unwrap_or_default();
-                sum_scritches += scritches;
-            }
-            sum_scritches
-        }
-
-        #[pg_extern]
-        fn sum_scritches_for_names_strict_optional_items(
-            dogs: Vec<Option<pgx::composite_type!("Dog")>>,
-        ) -> i32 {
-            // Gets resolved to:
-            let dogs: Vec<Option<PgHeapTuple<AllocatedByRust>>> = dogs;
-
-            let mut sum_scritches = 0;
-            for dog in dogs {
-                let dog = dog.unwrap();
-                let scritches: i32 = dog
-                    .get_by_name("scritches")
-                    .ok()
-                    .unwrap_or_default()
-                    .unwrap_or_default();
-                sum_scritches += scritches;
-            }
-            sum_scritches
-        }
-
-        #[pg_extern]
-        fn sum_scritches_for_names_default_optional_items(
-            dogs: pgx::default!(
-                Vec<Option<pgx::composite_type!("Dog")>>,
-                "ARRAY[ROW('Nami', 0)]::Dog[]"
-            ),
-        ) -> i32 {
-            // Gets resolved to:
-            let dogs: Vec<Option<PgHeapTuple<AllocatedByRust>>> = dogs;
-
-            let mut sum_scritches = 0;
-            for dog in dogs {
-                let dog = dog.unwrap();
-                let scritches: i32 = dog
-                    .get_by_name("scritches")
-                    .ok()
-                    .unwrap_or_default()
-                    .unwrap_or_default();
-                sum_scritches += scritches;
-            }
-            sum_scritches
-        }
-
-        #[pg_extern]
-        fn sum_scritches_for_names_optional_items(
-            dogs: Option<Vec<Option<pgx::composite_type!("Dog")>>>,
-        ) -> i32 {
-            // Gets resolved to:
-            let dogs: Option<Vec<Option<PgHeapTuple<AllocatedByRust>>>> = dogs;
-
-            let dogs = dogs.unwrap();
-            let mut sum_scritches = 0;
-            for dog in dogs {
-                let dog = dog.unwrap();
-                let scritches: i32 = dog
-                    .get_by_name("scritches")
-                    .ok()
-                    .unwrap_or_default()
-                    .unwrap_or_default();
-                sum_scritches += scritches;
-            }
-            sum_scritches
-        }
-    }
-
-    mod array {
-        use super::*;
-
-        #[pg_extern]
-        fn sum_scritches_for_names_array(
-            dogs: Option<pgx::Array<pgx::composite_type!("Dog")>>,
-        ) -> i32 {
-            // Gets resolved to:
-            let dogs: Option<pgx::Array<PgHeapTuple<AllocatedByRust>>> = dogs;
-
-            let dogs = dogs.unwrap();
-            let mut sum_scritches = 0;
-            for dog in dogs {
-                let dog = dog.unwrap();
-                let scritches: i32 = dog
-                    .get_by_name("scritches")
-                    .ok()
-                    .unwrap_or_default()
-                    .unwrap_or_default();
-                sum_scritches += scritches;
-            }
-            sum_scritches
-        }
-
-        #[pg_extern]
-        fn sum_scritches_for_names_array_default(
-            dogs: pgx::default!(
-                pgx::Array<pgx::composite_type!("Dog")>,
-                "ARRAY[ROW('Nami', 0)]::Dog[]"
-            ),
-        ) -> i32 {
-            // Gets resolved to:
-            let dogs: pgx::Array<PgHeapTuple<AllocatedByRust>> = dogs;
-
-            let mut sum_scritches = 0;
-            for dog in dogs {
-                let dog = dog.unwrap();
-                let scritches: i32 = dog
-                    .get_by_name("scritches")
-                    .ok()
-                    .unwrap_or_default()
-                    .unwrap_or_default();
-                sum_scritches += scritches;
-            }
-            sum_scritches
-        }
-
-        #[pg_extern]
-        fn sum_scritches_for_names_array_strict(
-            dogs: pgx::Array<pgx::composite_type!("Dog")>,
-        ) -> i32 {
-            // Gets resolved to:
-            let dogs: pgx::Array<PgHeapTuple<AllocatedByRust>> = dogs;
-
-            let mut sum_scritches = 0;
-            for dog in dogs {
-                let dog = dog.unwrap();
-                let scritches: i32 = dog
-                    .get_by_name("scritches")
-                    .ok()
-                    .unwrap_or_default()
-                    .unwrap_or_default();
-                sum_scritches += scritches;
-            }
-            sum_scritches
-        }
-    }
-}
-
-// As return types
-mod returning {
-    use super::*;
-
-    mod singleton {
-        use super::*;
-
-        #[pg_extern]
-        fn create_dog(_name: String, _scritches: i32) -> pgx::composite_type!("Dog") {
-            todo!()
-        }
-
-        #[pg_extern]
-        fn scritch(
-            maybe_dog: Option<::pgx::composite_type!("Dog")>,
-        ) -> Option<pgx::composite_type!("Dog")> {
-            // Gets resolved to:
-            let maybe_dog: Option<PgHeapTuple<AllocatedByRust>> = maybe_dog;
-
-            let maybe_dog = if let Some(mut dog) = maybe_dog {
-                dog.set_by_name("scritches", dog.get_by_name::<i32>("scritches").unwrap())
-                    .unwrap();
-                Some(dog)
-            } else {
-                None
-            };
-
-            maybe_dog
-        }
-
-        #[pg_extern]
-        fn scritch_strict(dog: ::pgx::composite_type!("Dog")) -> pgx::composite_type!("Dog") {
-            // Gets resolved to:
-            let mut dog: PgHeapTuple<AllocatedByRust> = dog;
-
-            dog.set_by_name("scritches", dog.get_by_name::<i32>("scritches").unwrap())
-                .unwrap();
-
-            dog
-        }
-    }
-
-    mod vec {
-        use super::*;
-
-        #[pg_extern]
-        fn scritch_all_vec(
-            maybe_dogs: Option<Vec<::pgx::composite_type!("Dog")>>,
-        ) -> Option<Vec<::pgx::composite_type!("Dog")>> {
-            // Gets resolved to:
-            let maybe_dogs: Option<Vec<PgHeapTuple<AllocatedByRust>>> = maybe_dogs;
-
-            let maybe_dogs = if let Some(mut dogs) = maybe_dogs {
-                for dog in dogs.iter_mut() {
-                    dog.set_by_name("scritches", dog.get_by_name::<i32>("scritches").unwrap())
-                        .unwrap();
-                }
-                Some(dogs)
-            } else {
-                None
-            };
-
-            maybe_dogs
-        }
-
-        #[pg_extern]
-        fn scritch_all_vec_strict(
-            dogs: Vec<::pgx::composite_type!("Dog")>,
-        ) -> Vec<::pgx::composite_type!("Dog")> {
-            // Gets resolved to:
-            let mut dogs: Vec<PgHeapTuple<AllocatedByRust>> = dogs;
-
-            for dog in dogs.iter_mut() {
-                dog.set_by_name("scritches", dog.get_by_name::<i32>("scritches").unwrap())
-                    .unwrap();
-            }
-            dogs
-        }
-
-        #[pg_extern]
-        fn scritch_all_vec_optional_items(
-            maybe_dogs: Option<Vec<Option<::pgx::composite_type!("Dog")>>>,
-        ) -> Option<Vec<Option<::pgx::composite_type!("Dog")>>> {
-            // Gets resolved to:
-            let maybe_dogs: Option<Vec<Option<PgHeapTuple<AllocatedByRust>>>> = maybe_dogs;
-
-            let maybe_dogs = if let Some(mut dogs) = maybe_dogs {
-                for dog in dogs.iter_mut() {
-                    if let Some(ref mut dog) = dog {
-                        dog.set_by_name("scritches", dog.get_by_name::<i32>("scritches").unwrap())
-                            .unwrap();
-                    }
-                }
-                Some(dogs)
-            } else {
-                None
-            };
-
-            maybe_dogs
-        }
-    }
-
-    // Returning VariadicArray/Array isn't supported, use a Vec.
-}
-
-// Just a compile test...
+```sql
+CREATE FUNCTION create_dog(name TEXT, scritches INT) RETURNS Dog
+    LANGUAGE SQL
+    STRICT RETURN ROW(name, scritches)::Dog;
+```
+*/
 #[pg_extern]
-fn exotic_signature(
-    _cats: pgx::default!(
-        Option<Vec<Option<::pgx::composite_type!("Cat")>>>,
-        "ARRAY[ROW('Sally', 0)]::Cat[]"
-    ),
-    _a_single_fish: pgx::default!(
-        Option<::pgx::composite_type!("Fish")>,
-        "ROW('Bob', 0)::Fish"
-    ),
-    _dogs: pgx::default!(
-        Option<::pgx::VariadicArray<::pgx::composite_type!("Dog")>>,
-        "ARRAY[ROW('Nami', 0), ROW('Brandy', 0)]::Dog[]"
-    ),
-) -> Option<Vec<::pgx::composite_type!("AnimalFriendshipEdge")>> {
-    unimplemented!()
+fn create_dog(name: String, scritches: i32) -> pgx::composite_type!(DOG_COMPOSITE_TYPE) {
+    let mut dog = PgHeapTuple::new_composite_type(DOG_COMPOSITE_TYPE).unwrap();
+    dog.set_by_name("name", name).unwrap();
+    dog.set_by_name("scritches", scritches).unwrap();
+    dog
+}
+
+/*
+Scritch the passed dog, then return it, roughly the equivalent of this SQL:
+
+```sql
+CREATE FUNCTION scritch_dog(dog Dog) RETURNS Dog
+    LANGUAGE SQL
+    STRICT RETURN ROW(dog.name, dog.scritches + 1)::Dog;
+```
+*/
+#[pg_extern]
+fn scritch_dog(mut dog: pgx::composite_type!(DOG_COMPOSITE_TYPE)) -> pgx::composite_type!(DOG_COMPOSITE_TYPE) {
+    if let Ok(scritches) = dog.get_by_name::<i32>("scritches") {
+        dog.set_by_name("scritches", scritches).unwrap();
+    }
+    dog
+}
+
+/*
+Make a cat and a dog friends, returning that relationship, roughly the same as:
+
+```sql
+CREATE FUNCTION make_friendship(dog Dog, cat Cat) RETURNS CatAndDogFriendship
+    LANGUAGE SQL
+    STRICT RETURN ROW(dog, cat)::CatAndDogFriendship;
+```
+This function primarily exists to demonstrate how to make some `extension_sql!()` "appear before"
+the function.
+*/
+#[pg_extern(requires = ["create_cat_and_dog_friendship"])]
+fn make_friendship(
+    dog: pgx::composite_type!(DOG_COMPOSITE_TYPE),
+    cat: pgx::composite_type!(CAT_COMPOSITE_TYPE)
+) -> pgx::composite_type!(CAT_AND_DOG_FRIENDSHIP_COMPOSITE_TYPE) {
+    let mut friendship = PgHeapTuple::new_composite_type(CAT_AND_DOG_FRIENDSHIP_COMPOSITE_TYPE).unwrap();
+    friendship.set_by_name("dog", dog).unwrap();
+    friendship.set_by_name("cat", cat).unwrap();
+    friendship
 }
 
 #[cfg(any(test, feature = "pg_test"))]
@@ -452,144 +145,35 @@ mod tests {
     use pgx::*;
 
     #[pg_test]
-    fn test_gets_name_field() {
-        let retval = Spi::get_one::<&str>(
-            "
-            SELECT gets_name_field(ROW('Nami', 0)::Dog)
-        ",
-        )
-        .expect("SQL select failed");
-        assert_eq!(retval, "Nami");
-    }
-
-    #[pg_test]
-    fn test_gets_name_field_default() {
-        let retval = Spi::get_one::<&str>(
-            "
-            SELECT gets_name_field_default()
-        ",
-        )
-        .expect("SQL select failed");
-        assert_eq!(retval, "Nami");
-    }
-
-    #[pg_test]
-    fn test_gets_name_field_strict() {
-        let retval = Spi::get_one::<&str>(
-            "
-            SELECT gets_name_field_strict(ROW('Nami', 0)::Dog)
-        ",
-        )
-        .expect("SQL select failed");
-        assert_eq!(retval, "Nami");
-    }
-
-    #[pg_test]
-    fn test_gets_name_field_variadic() {
-        let retval = Spi::get_one::<Vec<String>>("
-            SELECT gets_name_field_variadic(ROW('Nami', 1)::Dog, ROW('Brandy', 1)::Dog)
-        ").expect("SQL select failed");
-        assert_eq!(retval, vec!["Nami".to_string(), "Brandy".to_string()]);
-    }
-
-    #[pg_test]
-    fn test_sum_scritches_for_names() {
-        let retval = Spi::get_one::<i32>("
-            SELECT sum_scritches_for_names(ARRAY[ROW('Nami', 1), ROW('Brandy', 42)]::Dog[])
-        ").expect("SQL select failed");
-        assert_eq!(retval, 43);
-    }
-
-    #[pg_test]
-    fn test_sum_scritches_for_names_default() {
-        let retval = Spi::get_one::<i32>("
-            SELECT sum_scritches_for_names_default()
-        ").expect("SQL select failed");
-        assert_eq!(retval, 0);
-    }
-
-    #[pg_test]
-    fn test_sum_scritches_for_names_strict() {
-        let retval = Spi::get_one::<i32>("
-            SELECT sum_scritches_for_names_strict(ARRAY[ROW('Nami', 1), ROW('Brandy', 42)]::Dog[])
-        ").expect("SQL select failed");
-        assert_eq!(retval, 43);
-    }
-
-    #[pg_test]
-    fn test_sum_scritches_for_names_strict_optional_items() {
-        let retval = Spi::get_one::<i32>("
-            SELECT sum_scritches_for_names_strict(ARRAY[ROW('Nami', 1), ROW('Brandy', 42)]::Dog[])
-        ").expect("SQL select failed");
-        assert_eq!(retval, 43);
-    }
-
-    #[pg_test]
-    fn test_sum_scritches_for_names_default_optional_items() {
-        let retval = Spi::get_one::<i32>("
-            SELECT sum_scritches_for_names_default_optional_items()
-        ").expect("SQL select failed");
-        assert_eq!(retval, 0);
-    }
-
-    #[pg_test]
-    fn test_sum_scritches_for_names_optional_items() {
-        let retval = Spi::get_one::<i32>("
-            SELECT sum_scritches_for_names_optional_items(ARRAY[ROW('Nami', 1), ROW('Brandy', 42)]::Dog[])
-        ").expect("SQL select failed");
-        assert_eq!(retval, 0);
-    }
-
-    #[pg_test]
-    fn test_sum_scritches_for_names_array() {
-        let retval = Spi::get_one::<i32>("
-            SELECT sum_scritches_for_names_array(ARRAY[ROW('Nami', 1), ROW('Brandy', 42)]::Dog[])
-        ").expect("SQL select failed");
-        assert_eq!(retval, 43);
-    }
-
-    #[pg_test]
-    fn test_sum_scritches_for_names_array_default() {
-        let retval = Spi::get_one::<i32>("
-            SELECT sum_scritches_for_names_array_default()
-        ").expect("SQL select failed");
-        assert_eq!(retval, 0);
-    }
-
-    #[pg_test]
-    fn test_sum_scritches_for_names_array_strict() {
-        let retval = Spi::get_one::<i32>("
-            SELECT sum_scritches_for_names_array_strict(ARRAY[ROW('Nami', 1), ROW('Brandy', 42)]::Dog[])
-        ").expect("SQL select failed");
-        assert_eq!(retval, 43);
-    }
-
-    #[pg_test]
     fn test_create_dog() {
-        let retval = Spi::get_one::<PgHeapTuple<'_, AllocatedByRust>>("
-            SELECT create_dog('Nami', 1)
+        let retval = Spi::get_one::<PgHeapTuple<AllocatedByRust>>("\
+            SELECT create_dog('Nami', 0)
         ").expect("SQL select failed");
-        assert_eq!(retval.get_by_name("name").unwrap(), Some("Nami"));
-        assert_eq!(retval.get_by_name("scritches").unwrap(), Some(1));
+        assert_eq!(retval.get_by_name::<&str>("name").unwrap().unwrap(), "Nami");
+        assert_eq!(retval.get_by_name::<i32>("scritches").unwrap().unwrap(), 0);
     }
 
     #[pg_test]
-    fn test_scritch() {
-        let retval = Spi::get_one::<PgHeapTuple<'_, AllocatedByRust>>("
-            SELECT scritch(ROW('Nami', 1))
+    fn test_scritch_dog() {
+        let retval = Spi::get_one::<PgHeapTuple<AllocatedByRust>>("\
+            SELECT scritch_dog(ROW('Nami', 1)::Dog)
         ").expect("SQL select failed");
-        assert_eq!(retval.get_by_name("name").unwrap(), Some("Nami"));
-        assert_eq!(retval.get_by_name("scritches").unwrap(), Some(2));
+        assert_eq!(retval.get_by_name::<&str>("name").unwrap().unwrap(), "Nami");
+        assert_eq!(retval.get_by_name::<i32>("scritches").unwrap().unwrap(), 1);
     }
 
     #[pg_test]
-    fn test_scritch_strict() {
-        let retval = Spi::get_one::<PgHeapTuple<'_, AllocatedByRust>>("
-            SELECT scritch_strict(ROW('Nami', 1))
+    fn test_make_friendship() {
+        let friendship = Spi::get_one::<PgHeapTuple<AllocatedByRust>>("\
+            SELECT make_friendship(ROW('Nami', 0)::Dog, ROW('Sally', 0)::Cat)
         ").expect("SQL select failed");
-        assert_eq!(retval.get_by_name("name").unwrap(), Some("Nami"));
-        assert_eq!(retval.get_by_name("scritches").unwrap(), Some(2));
+        let dog: PgHeapTuple<AllocatedByRust> = friendship.get_by_name("dog").unwrap().unwrap();
+        assert_eq!(dog.get_by_name::<&str>("name").unwrap().unwrap(), "Nami");
+
+        let cat: PgHeapTuple<AllocatedByRust> = friendship.get_by_name("cat").unwrap().unwrap();
+        assert_eq!(cat.get_by_name::<&str>("name").unwrap().unwrap(), "Sally");
     }
+
 }
 
 #[cfg(test)]
